made some changes to sum in vector and indirect

+++ sum_vector.cpp
@@ -21,7 +21,7 @@ int64_t
 sum(int64_t N, uint64_t A[])
 {
    printf(" inside sum_vector perform_sum, N=%lld \n", N);
-   int sum = 0;
+   uint64_t sum = 0;
    for(int i = 0; i < N; i ++){
       sum += A[i];
    }
